@@ -13,6 +13,8 @@
   "dependencies": {
     "@ethersproject/contracts": "^5.8.0",
     "@ethersproject/providers": "^5.8.0",
+    "@tanstack/react-query": "^5.83.0",
+    "@web3modal/wagmi": "^5.1.11",
     "body-parser": "^1.20.3",
     "chalk": "^2.4.2",
     "cookie-parser": "^1.4.7",
@@ -30,19 +32,21 @@
     "lodash": "^4.17.21",
     "prop-types": "^15.8.1",
     "query-string": "^7.1.3",
-    "react": "^16.14.0",
-    "react-bootstrap": "^1.6.7",
-    "react-dom": "^16.14.0",
+    "react": "^18.3.1",
+    "react-bootstrap": "^1.6.8",
+    "react-dom": "^18.3.1",
     "react-markdown": "^6.0.3",
-    "react-markdown-it": "^1.0.2",
     "react-password-strength-bar": "^0.3.5",
     "react-player": "^2.16.1",
+    "react-router-dom": "^7.7.0",
     "react-scroll": "^1.9.3",
     "react-webcam": "^5.2.4",
     "require-or-mock": "^0.2.1",
     "simple-json-db": "^1.3.0",
     "superagent": "^4.1.0",
     "video-react": "^0.14.1",
+    "viem": "^2.32.1",
+    "wagmi": "^2.15.7",
     "web3modal": "^1.9.12"
   },
   "devDependencies": {
