@@ -1,6 +1,6 @@
+import React from 'react'
 import ls from 'local-storage'
 
-// eslint-disable-next-line no-undef
 export default class Common extends React.Component {
 
   constructor(props) {
