@@ -1,6 +1,5 @@
-// eslint-disable-next-line no-undef
-const {Row, Col} = ReactBootstrap
-
+import React from 'react'
+import { Row, Col } from 'react-bootstrap'
 import Ab from '../Ab'
 
 export default class CenteredImage extends React.Component {
